requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

# azure-functions
azure-functions
azure-storage-blob
numpy
opencv-python-headless
easyocr
torch
torchvision